+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Version latest (0.18.0rc?, 2019-12-05?)
 - enh: cost analyze: after sending a verification email, use click.pause instead of click.prompt
 - enh: cost analyze: only save "last-used email" after the email verification is complete
 - feat: cost: save last-used profile in `~/.isitfit/last_profile.txt`
+- feat: add colors to profile names in prompt
 
 
 Version 0.18.6 (2019-12-11)